# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "FlowGraph API",
    "version": "0.1.0",
    "description": "Visual set-planning for DJs. Tracks are nodes, transitions are directed edges, and a set is a versioned path through the graph.\n\nEvery endpoint is scoped to the caller's workspace, derived from the session cookie. Clients never choose a workspace."
  },
  "paths": {
...
    "/v1/graphs/{graphId}/nodes/{nodeId}": {
      "delete": {
        "responses": {
          "200": {
            "description": "Node removed."
          },
          "401": {
            "description": "No valid session.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "statusCode",
                    "message",
                    "error",
                    "timestamp",
                    "path",
                    "requestId"
                  ],
                  "properties": {
                    "statusCode": {
                      "type": "integer",
                      "minimum": -9007199254740991,
                      "maximum": 9007199254740991
                    },
                    "message": {
                      "type": "string"
                    },
                    "error": {
                      "type": "string"
                    },
                    "timestamp": {
                      "type": "string",
                      "format": "date-time",
                      "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$"
                    },
                    "path": {
                      "type": "string"
                    },
                    "requestId": {
                      "type": "string"
                    },
                    "details": {
                      "type": "object",
                      "required": [
                        "errors",
                        "errorCount"
                      ],
                      "properties": {
                        "errors": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "required": [
                              "path",
                              "message",
                              "code"
                            ],
                            "properties": {
                              "path": {
                                "type": "array",
                                "items": {
                                  "anyOf": [
                                    {
                                      "type": "string"
                                    },
                                    {
                                      "type": "number"
                                    }
                                  ]
                                }
                              },
                              "message": {
                                "type": "string"
                              },
                              "code": {
                                "type": "string"
                              },
                              "expected": {
                                "type": "string"
                              }
                            },
                            "additionalProperties": false
                          }
                        },
                        "errorCount": {
                          "type": "integer",
                          "minimum": -9007199254740991,
                          "maximum": 9007199254740991
                        }
                      },
                      "additionalProperties": false
                    }
                  },
                  "additionalProperties": false
                }
              }
            }
          },
          "403": {
            "description": "No workspace.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "statusCode",
                    "message",
                    "error",
                    "timestamp",
                    "path",
                    "requestId"
                  ],
                  "properties": {
                    "statusCode": {
                      "type": "integer",
                      "minimum": -9007199254740991,
                      "maximum": 9007199254740991
                    },
                    "message": {
                      "type": "string"
                    },
                    "error": {
                      "type": "string"
                    },
                    "timestamp": {
                      "type": "string",
                      "format": "date-time",
                      "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$"
                    },
                    "path": {
                      "type": "string"
                    },
                    "requestId": {
                      "type": "string"
                    },
                    "details": {
                      "type": "object",
                      "required": [
                        "errors",
                        "errorCount"
                      ],
                      "properties": {
                        "errors": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "required": [
                              "path",
                              "message",
                              "code"
                            ],
                            "properties": {
                              "path": {
                                "type": "array",
                                "items": {
                                  "anyOf": [
                                    {
                                      "type": "string"
                                    },
                                    {
                                      "type": "number"
                                    }
                                  ]
                                }
                              },
                              "message": {
                                "type": "string"
                              },
                              "code": {
                                "type": "string"
                              },
                              "expected": {
                                "type": "string"
                              }
                            },
                            "additionalProperties": false
                          }
                        },
                        "errorCount": {
                          "type": "integer",
                          "minimum": -9007199254740991,
                          "maximum": 9007199254740991
                        }
                      },
                      "additionalProperties": false
                    }
                  },
                  "additionalProperties": false
                }
              }
            }
          },
          "404": {
            "description": "No such node.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "statusCode",
                    "message",
                    "error",
                    "timestamp",
                    "path",
                    "requestId"
                  ],
                  "properties": {
                    "statusCode": {
                      "type": "integer",
                      "minimum": -9007199254740991,
                      "maximum": 9007199254740991
                    },
                    "message": {
                      "type": "string"
                    },
                    "error": {
                      "type": "string"
                    },
                    "timestamp": {
                      "type": "string",
                      "format": "date-time",
                      "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$"
                    },
                    "path": {
                      "type": "string"
                    },
                    "requestId": {
                      "type": "string"
                    },
                    "details": {
                      "type": "object",
                      "required": [
                        "errors",
                        "errorCount"
                      ],
                      "properties": {
                        "errors": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "required": [
                              "path",
                              "message",
                              "code"
                            ],
                            "properties": {
                              "path": {
                                "type": "array",
                                "items": {
                                  "anyOf": [
                                    {
                                      "type": "string"
                                    },
                                    {
                                      "type": "number"
                                    }
                                  ]
                                }
                              },
                              "message": {
                                "type": "string"
                              },
                              "code": {
                                "type": "string"
                              },
                              "expected": {
                                "type": "string"
                              }
                            },
                            "additionalProperties": false
                          }
                        },
                        "errorCount": {
                          "type": "integer",
                          "minimum": -9007199254740991,
                          "maximum": 9007199254740991
                        }
                      },
                      "additionalProperties": false
                    }
                  },
                  "additionalProperties": false
                }
              }
            }
          }
        },
        "tags": [
          "Graphs"
        ],
        "summary": "Remove a track from the graph",
        "description": "Delete-from-graph, not delete-from-library — the track stays in the workspace and on any other canvas.",
        "parameters": [
          {
            "name": "nodeId",
            "in": "path",
            "required": true,
            "description": "Node UUID.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "graphId",
            "in": "path",
            "required": true,
            "description": "Graph UUID.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "sessionCookie": []
          }
        ]
      }
    },
...
  },
...
  "servers": [
    {
      "url": "http://127.0.0.1:4000",
      "description": "Local development"
    }
  ],
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "description": "Session cookie issued by /api/auth/sign-in/email. State-changing auth routes additionally require an Origin header from AUTH_TRUSTED_ORIGINS (CSRF protection).",
        "name": "better-auth.session_token",
        "in": "cookie"
      }
    }
  }
}